+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Faiadashu
 * Added displaying units in numeric answer input
 * Added questionnaire-unitOption extension support to make ability choose quantity units in numeric answers
+* Fixed drop down style for numeric answer input
 
 ## 0.9.0-alpha.1
 **[Breaking]** This alpha release is beginning to use Flutter 3.10.0 and fhir 0.11.0.